requirements.txt
python-telegram-bot==20.6
google-generativeai==0.3.2
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.2.1
